Show cooldowns on the earn screen
