+++ app/src/main/res/layout/activity_earn_info.xml
@@ -89,6 +89,13 @@
                     android:text="@string/earn_delegation_button" />
             </LinearLayout>
 
+            <LinearLayout
+                android:id="@+id/cooldown_list_container"
+                android:layout_width="match_parent"
+                android:layout_height="wrap_content"
+                android:layout_marginTop="12dp"
+                android:orientation="vertical" />
+
             <Space
                 android:layout_width="0dp"
                 android:layout_height="20dp" />
