+++ app/src/main/res/layout/item_cooldown.xml
@@ -1,0 +1,8 @@
+<?xml version="1.0" encoding="utf-8"?>
+<com.concordium.wallet.ui.bakerdelegation.common.CooldownView xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:tools="http://schemas.android.com/tools"
+    style="@style/CryptoX_Container_TextCard"
+    android:layout_width="match_parent"
+    android:layout_height="wrap_content"
+    android:layout_marginTop="8dp"
+    tools:viewBindingIgnore="true" />
